requirements.txt
django
openai-whisper
torch
pymongo
python-dotenv
pydub
soundfile
numpy
fastapi
uvicorn
